No json in base card builder
