+++ touchvie_front/src/main/java/com/touchvie/touchvie_front/builders/CardDetailBuilder.java
@@ -1,5 +1,7 @@
 package com.touchvie.touchvie_front.builders;
 
+import android.content.Context;
+
 /**
  * Class to build a card detail without a configuration json.
  */
@@ -8,7 +10,8 @@ public class CardDetailBuilder extends BaseCardDetailBuilder<CardDetailBuilder> 
     /**
      * Default constructor
      */
-    public CardDetailBuilder() {
+    public CardDetailBuilder(Context context) {
+        super(context);
     }
