+++ touchvie_front/src/main/java/com/touchvie/touchvie_front/builders/CardDetailJson.java
@@ -1,6 +1,16 @@
 package com.touchvie.touchvie_front.builders;
 
+import android.content.Context;
+import android.content.res.Resources;
+
+import com.google.gson.GsonBuilder;
+import com.touchvie.backend.DataConfig;
+import com.touchvie.touchvie_front.R;
+
 import org.json.JSONObject;
+
+import java.io.InputStream;
+import java.util.HashMap;
 
 /**
  * Class to build a card detail with a configuration json.
@@ -10,7 +20,8 @@ public class CardDetailJson extends BaseCardDetailBuilder<CardDetailJson> {
     /**
      * Default constructor
      */
-    public CardDetailJson() {
+    public CardDetailJson(Context context) {
+        super(context);
     }
 
     /**
@@ -26,7 +37,39 @@ public class CardDetailJson extends BaseCardDetailBuilder<CardDetailJson> {
      */
     public CardDetailJson loadDataConfig(JSONObject dataConfig) {
         // TODO implement here
+        requestDataConfig();
         return getThis();
+    }
+
+    /**
+     * FOR TESTING PURPOSES
+     */
+
+    private void requestDataConfig() {
+        String jsonString = null;
+        try {
+            Resources res = context.getResources();
+            InputStream in_s = res.openRawResource(R.raw.dataconfig);
+
+            byte[] b = new byte[in_s.available()];
+            in_s.read(b);
+            jsonString = new String(b);
+        } catch (Exception e) {
+            return;
+            //jsonString = new String(response.data, HttpHeaderParser.parseCharset(response.headers));
+        }
+
+        DataConfig dataConfig = new GsonBuilder().create().fromJson(jsonString, DataConfig.class);
+        if (dataConfig == null) {
+            dataConfig = new DataConfig();
+        }
+
+        idSection = new HashMap<>();
+        if (dataConfig.getSections() != null && dataConfig.getSections().length > 0) {
+            for (int i = 0; i < dataConfig.getSections().length; i++) {
+                idSection.put((dataConfig.getSections()[i]).getTitle(), dataConfig.getSections()[i]);
+            }
+        }
     }
 
     @Override
